+++ cxroots/root_finding.py
@@ -1,16 +1,19 @@
-from __future__ import division
-
 import functools
 import logging
 import warnings
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, Generator, Iterable, List, NamedTuple, Optional
 
 import numpy as np
 from numpydoc.docscrape import FunctionDoc
 from rich.progress import BarColumn, Progress, TextColumn
 
+from .contour_interface import ContourABC
 from .iterative_methods import iterate_to_root
+from .root_approximation import approximate_roots
 from .root_counting import RootError
 from .root_result import RootResult
+from .types import AnalyticFunc, IntegrationMethod
 from .util import NumberOfRootsChanged, update_docstring
 
 
@@ -23,12 +26,12 @@ class CountCalls:
     Count how many times a given function is called.
     """
 
-    def __init__(self, func):
+    def __init__(self, func: Callable[[Any], Any]):
         self.func = func
         self.calls = 0
         self.points = 0
 
-    def __call__(self, z):
+    def __call__(self, z: Any) -> Any:
         self.calls += 1
         if hasattr(z, "__len__"):
             self.points += len(z)
@@ -37,27 +40,46 @@ class CountCalls:
         return self.func(z)
 
 
+@dataclass
+class ContourData:
+    contour: ContourABC
+    num_roots: int
+
+
+def make_contour_data(
+    contours: List[ContourABC], num_roots: Dict[ContourABC, int]
+) -> List[ContourData]:
+    return [ContourData(contour=c, num_roots=num_roots[c]) for c in contours]
+
+
+class RootFinderState(NamedTuple):
+    roots: List[complex]
+    multiplicities: List[int]
+    contour_data: List[ContourData]
+    num_remaining_roots: int
+
+
 def find_roots_gen(
-    original_contour,
-    f,
-    df=None,
-    guess_roots=[],
-    guess_roots_symmetry=None,
-    newton_step_tol=1e-14,
-    attempt_best_iter=True,
-    newton_max_iter=50,
-    root_err_tol=1e-10,
-    abs_tol=0,
-    rel_tol=1e-12,
-    integer_tol=0.1,
-    int_abs_tol=0.07,
-    M=5,  # noqa: N803
-    err_stop=1e-10,
-    int_method="quad",
-    div_min=3,
-    div_max=15,
-    df_approx_order=2,
-):
+    original_contour: ContourABC,
+    f: AnalyticFunc,
+    df: Optional[AnalyticFunc] = None,
+    guess_roots: Iterable[complex] = [],
+    guess_roots_symmetry: Optional[Callable[[complex], Iterable[complex]]] = None,
+    newton_step_tol: float = 1e-14,
+    refine_roots_beyond_tol: bool = True,
+    newton_max_iter: int = 50,
+    root_err_tol: float = 1e-10,
+    abs_tol: float = 1.49e-08,
+    rel_tol: float = 1.49e-08,
+    integer_tol: float = 0.1,
+    int_abs_tol: float = 0.07,
+    M: int = 5,  # noqa: N803
+    err_stop: float = 1e-10,
+    int_method: IntegrationMethod = "quad",
+    div_min: int = 3,
+    div_max: int = 15,
+    df_approx_order: int = 2,
+) -> Generator[RootFinderState, None, None]:
     """
     A generator which at each step takes a contour and either finds all
     the zeros of f within it or subdivides it further.  Based on the
@@ -86,8 +108,8 @@ def find_roots_gen(
         The required accuracy of the root.  The iterative method used to
         give a final value for each root will exit if the step size, dx,
         between sucessive iterations satisfies abs(dx) < newton_step_tol
-        and iterBestAttempt is False.
-    attempt_best_iter : bool, optional
+        and refine_roots_beyond_tol is False.
+    refine_roots_beyond_tol : bool, optional
         If True then the iterative method used to refine the roots will
         exit when error of the previous iteration, x0, was at least as
         good as the current iteration, x, in the sense that
@@ -152,11 +174,11 @@ def find_roots_gen(
 
     Yields
     ------
-    list
+    list[complex]
         Roots of f(z) within the contour original_contour
-    list
+    list[int]
         Multiplicites of roots
-    deque
+    list[(contour, roots_in_contour)]
         The contours which still contain roots
     int
         Remaining number of roots to be found within the contour
@@ -182,76 +204,66 @@ def find_roots_gen(
         "int_method": int_method,
     }
 
+    logger = logging.getLogger(__name__)
+    roots: List[complex] = []
+    multiplicities: List[int] = []
+    failed_contours: List[ContourABC] = []
+    contours: List[ContourABC] = []
+    num_roots: Dict[ContourABC, int] = {}
+
     try:
         # compute the total number of zeros, including multiplicities, within the
         # originally given contour
-        original_contour._num_roots = original_contour.count_roots(**count_kwargs)
+        num_roots[original_contour] = original_contour.count_roots(**count_kwargs)
+        contours.append(original_contour)
     except RuntimeError:
         raise RuntimeError(
-            """
-            Integration along the initial contour has failed.
+            """Integration along the initial contour has failed.
             There is likely a root on or close to the initial contour.
             Try changing the initial contour, if possible."""
         )
 
-    logger = logging.getLogger(__name__)
-    logger.info(
-        "Counted "
-        + str(original_contour._num_roots)
-        + " roots (including multiplicities) within the original contour"
-    )
-
-    roots = []
-    multiplicities = []
-    failed_contours = []
-    contours = []
-    contours.append(original_contour)
-
-    def subdivide(parent_contour):
+    def subdivide(parent: ContourABC) -> None:
         """Given a contour, parent_contour, subdivide it into multiple contours."""
-        logger.info("Subdividing " + str(parent_contour))
+        logger.info("Subdividing " + str(parent))
 
-        num_roots = None
-        for subcontours in parent_contour.subdivisions():
+        for subcontours in parent.subdivisions():
             # if a contour has already been used and caused an error then skip it
             failed_contour_desc = list(map(str, failed_contours))
-            if np.any(
-                [
-                    contour_desc in failed_contour_desc
-                    for contour_desc in list(map(str, subcontours))
-                ]
+            if any(
+                contour_desc in failed_contour_desc
+                for contour_desc in list(map(str, subcontours))
             ):
                 continue
 
-            # if a contour is near to a known root then skip it
+            # if a contour is near to a known root then skip this subdivision
             for root in roots:
-                if np.any(
-                    np.abs([contour.distance(root) for contour in subcontours]) < 0.01
-                ):
+                if any(contour.distance(root) < 0.01 for contour in subcontours):
                     continue
 
             try:
-                num_roots = [
+                subcontour_num_roots = [
                     contour.count_roots(**count_kwargs)
                     for contour in np.array(subcontours)
                 ]
-                while parent_contour._num_roots != sum(num_roots):
+                while (
+                    num_roots[parent] != sum(subcontour_num_roots)
+                    and int_abs_tol > 0.01
+                ):
                     logger.warning(
                         "Number of roots in sub contours not adding up to parent "
-                        "contour.  Recomputing number of roots in parent and child "
+                        "contour. Recomputing number of roots in parent and child "
                         f"contours with int_abs_tol={0.5 * int_abs_tol}"
                     )
                     temp_count_kwargs = count_kwargs.copy()
                     temp_count_kwargs["int_abs_tol"] *= 0.5
-                    parent_contour._num_roots = parent_contour.count_roots(
-                        **temp_count_kwargs
-                    )
-                    num_roots = [
+                    num_roots[parent] = parent.count_roots(**temp_count_kwargs)
+                    subcontour_num_roots = [
                         contour.count_roots(**temp_count_kwargs)
                         for contour in np.array(subcontours)
                     ]
 
-                if parent_contour._num_roots == sum(num_roots):
+                if num_roots[parent] == sum(subcontour_num_roots):
                     break
 
             except RootError:
@@ -261,57 +273,54 @@ def find_roots_gen(
                 # the subdividing line at a different point along the division axis
                 logger.warning(
                     "RootError encountered when subdivding "
-                    + str(parent_contour)
+                    + str(parent)
                     + " into:\n"
                     + str(subcontours[0])
                     + "\n"
-                    + subcontours[1]
+                    + str(subcontours[1])
                 )
                 continue
-
-        if num_roots is None or parent_contour._num_roots != sum(num_roots):
-            # The list of subdivisions has been exhaused and still the number of
-            # enclosed zeros does not add up
+        else:
+            # The list of possible subdivisions has been exhaused
             raise RuntimeError(
-                """Unable to subdivide contour:
-                \t%s
+                f"""Unable to subdivide contour:
+                \t{parent}
                 """
-                % parent_contour
             )
 
-        # record number of roots within each sub-contour
-        for i, contour in enumerate(subcontours):
-            contour._num_roots = num_roots[i]
+        # add sub-contours with roots to the list of contours to find the roots in
+        for n_roots, c in zip(subcontour_num_roots, subcontours):
+            if n_roots != 0:
+                contours.append(c)
+                num_roots[c] = n_roots
 
-        # add these sub-contours to the list of contours to find the roots in
-        contours.extend(
-            [contour for i, contour in enumerate(subcontours) if num_roots[i] != 0]
-        )
-
-    def remove_siblings_children(contour):
+    def reverse_subdivision(contour: ContourABC) -> None:
         """
-        Remove the contour and all its siblings and children from the
-        list of contours to be examined/subdivided.
+        Remove the contour and all its siblings and children from the list of contours
+        to be examined/subdivided and put its parent back in the list to be subdivided
         """
-        try:
-            contours.remove(contour)
-        except ValueError:
-            pass
+        if contour.parent is None:
+            raise ValueError("Contour has no parent so subdivision cannot be reversed")
 
-        # get sibling and child contours
-        # siblings:
-        relations = contour._parent._children
-        relations.remove(contour)
+        relations = []
+        if contour.parent.children is not None:
+            # The contour has siblings
+            relations.extend([c for c in contour.parent.children if c != contour])
 
-        # children:
-        if hasattr(contour, "_children"):
-            relations.extend(contour._children)
+        if contour.children is not None:
+            relations.extend(contour.children)
 
-        # interate over all relations
-        for relation in relations:
-            remove_siblings_children(relation)
+        for c in contours.copy():
+            if c in relations:
+                contours.remove(c)
 
-    def add_root(root, multiplicity):
+        # put the parent contour back into the list of contours to subdivide again
+        contours.append(contour.parent)
+
+        # do not use this contour again
+        failed_contours.append(contour)
+
+    def add_root(root: complex, multiplicity: int) -> None:
         # check that the root we have found is distinct from the ones we already have
         if not roots or np.all(abs(np.array(roots) - root) > newton_step_tol):
             # add the root to the list if it is within the original contour
@@ -335,51 +344,45 @@ def find_roots_gen(
                     # first check that x0 is distinct from the roots we already have
                     if np.all(abs(np.array(roots) - x0) > newton_step_tol):
                         logger.info(str(root) + " is a root so checking " + str(x0))
-                        root = iterate_to_root(
+                        new_root = iterate_to_root(
                             x0,
                             f,
                             df,
                             newton_step_tol,
                             root_err_tol,
                             newton_max_iter,
-                            attempt_best_iter,
+                            refine_roots_beyond_tol,
                         )
-                        if root is not None:
-                            contours.append(Circle(root, 1e-3))
-                            contours[-1]._num_roots = contours[-1].count_roots(
-                                **count_kwargs
-                            )
+                        if new_root is not None:
+                            circle = Circle(new_root, 1e-3)
+                            contours.append(circle)
+                            num_roots[circle] = circle.count_roots(**count_kwargs)
         else:
             logger.debug("Already recorded root " + str(root))
 
     # Add contours surrounding known roots so that they will be checked
     for root in guess_roots:
-        contours.append(Circle(root, 1e-3))
-        contours[-1]._num_roots = contours[-1].count_roots(**count_kwargs)
+        circle = Circle(root, 1e-3)
+        num_roots[circle] = circle.count_roots(**count_kwargs)
+        contours.append(circle)
 
     while contours:
         # yield the initial state here so that the animation in demo_find_roots shows
         # the first frame
-        num_found_roots = sum(
-            int(round(multiplicity.real))
-            for root, multiplicity in zip(roots, multiplicities)
+        num_found_roots = sum(multiplicities)
+        remaining_roots = num_roots[original_contour] - num_found_roots
+        yield RootFinderState(
+            roots,
+            multiplicities,
+            make_contour_data(contours, num_roots),
+            remaining_roots,
         )
-        remaining_roots = original_contour._num_roots - num_found_roots
-        yield roots, multiplicities, contours, remaining_roots
         contour = contours.pop()
 
         # if a known root is too near to this contour then reverse the subdivision that
         # created it
         if np.any([contour.distance(root) < newton_step_tol for root in roots]):
-            # remove the contour together with its children and siblings
-            remove_siblings_children(contour)
-
-            # put the parent contour back into the list of contours to be subdivided
-            # again
-            contours.append(contour._parent)
-
-            # do not use this contour again
-            failed_contours.append(contour)
+            reverse_subdivision(contour)
             continue
 
         # if the contour is smaller than the newton_step_tol then just assume that
@@ -392,7 +395,7 @@ def find_roots_gen(
                 newton_step_tol,
                 root_err_tol,
                 newton_max_iter,
-                attempt_best_iter,
+                refine_roots_beyond_tol,
             )
             if (
                 root is None
@@ -404,28 +407,26 @@ def find_roots_gen(
             warnings.warn(
                 "The area of this contour is smaller than newton_step_tol. Try "
                 f"increasing root_tol. The point z = {root.real} + {root.imag}i has "
-                f"been recorded as a root of multiplicity {contour._num_roots}. "
+                f"been recorded as a root of multiplicity {num_roots[contour]}. "
                 f"The error |f(z)| = {abs(f(root))}"
             )
-            add_root(root, contour._num_roots)
+            add_root(root, num_roots[contour])
             continue
 
         # if all the roots within the contour have been located then continue to the
         # next contour
         num_known_roots = sum(
-            [
-                int(round(multiplicity.real))
-                for root, multiplicity in zip(roots, multiplicities)
-                if contour.contains(root)
-            ]
+            multiplicity
+            for root, multiplicity in zip(roots, multiplicities)
+            if contour.contains(root)
         )
-        if contour._num_roots == num_known_roots:
+        if num_roots[contour] == num_known_roots:
             continue
 
         # if there are too many roots within the contour then subdivide or
         # if there are any known roots within the contour then also subdivide
         # (so as not to waste time re-approximating these roots)
-        if num_known_roots > 0 or contour._num_roots > M:
+        if num_known_roots > 0 or num_roots[contour] > M:
             subdivide(contour)
             continue
 
@@ -433,40 +434,39 @@ def find_roots_gen(
         if int_method == "romb":
             # Check to see if the number of roots has changed after new values of f
             # have been sampled
+            new_count_kwargs = count_kwargs.copy()
+
             def callback(integral, err, num_div):
-                if num_div > contour._num_divisions_for_N:
+                if num_div > new_count_kwargs["div_min"]:
                     logger.info(
                         "Checking root count using the newly sampled values of f"
                     )
-                    new_num_roots = contour.count_roots(
-                        f,
-                        df,
-                        int_abs_tol=int_abs_tol,
-                        integer_tol=integer_tol,
-                        div_min=num_div,
-                        div_max=div_max,
-                        df_approx_order=df_approx_order,
-                        int_method=int_method,
-                    )
+                    new_count_kwargs["div_min"] = num_div
+                    new_num_roots = contour.count_roots(**new_count_kwargs)
 
-                    if new_num_roots != contour._num_roots:
-                        logger.info("N has been recalculated using more samples of f")
-                        original_num_roots = contour._num_roots
-                        contour._num_roots = new_num_roots
+                    if new_num_roots != num_roots[contour]:
+                        logger.info(
+                            "Number of roots has been recalculated using more samples"
+                            " of f"
+                        )
+                        original_num_roots = num_roots[contour]
+                        num_roots[contour] = new_num_roots
                         raise NumberOfRootsChanged(
                             "The additional function evaluations of f taken while "
-                            "approximating the roots within the contour have been "
+                            "approximating the roots within the contour have "
                             "shown that the number of roots of f within the contour "
-                            f"is {new_num_roots} rather than the supplied "
-                            f"{original_num_roots}."
+                            f"is {new_num_roots} instead of {original_num_roots}."
                         )
 
         else:
-            callback = None
+
+            def callback(integral, err, num_div):
+                pass
 
         try:
-            approx_roots, approx_multiplicities = contour.approximate_roots(
-                contour._num_roots,
+            approx_roots, approx_multiplicities = approximate_roots(
+                contour,
+                num_roots[contour],
                 f,
                 df,
                 abs_tol=abs_tol,
@@ -481,7 +481,7 @@ def find_roots_gen(
             )
         except NumberOfRootsChanged:
             logger.debug("The number of roots within the contour has been reevaluated")
-            if contour._num_roots > M:
+            if num_roots[contour] > M:
                 subdivide(contour)
             else:
                 contours.append(contour)
@@ -507,7 +507,7 @@ def find_roots_gen(
                 newton_step_tol,
                 root_err_tol,
                 newton_max_iter,
-                attempt_best_iter,
+                refine_roots_beyond_tol,
             )
 
             if root is None or abs(f(approx_root)) < abs(f(root)):
@@ -520,14 +520,17 @@ def find_roots_gen(
                     > integer_tol
                 ):
                     # the computed multiplicity might be unreliable so make a contour
-                    # focused on that point instead
-                    if hasattr(contour, "_shrinking_radius"):
-                        contour._shrinking_radius *= 0.5
-                        contours.append(Circle(root, contour._shrinking_radius))
-                    else:
-                        contours.append(Circle(root, 1e-3))
-                        contours[-1]._shrinking_radius = 1e-3
-                    contours[-1]._num_roots = contours[-1].count_roots(**count_kwargs)
+                    # focused on that point instead so that the multiplicity will be
+                    # recomputed. If the current contour is already centered on that
+                    # point then make one with a smaller radius
+                    radius = (
+                        min(contour.radius * 0.1, 1e-3)
+                        if isinstance(contour, Circle) and contour.center == root
+                        else 1e-3
+                    )
+                    circle = Circle(root, radius)
+                    contours.append(circle)
+                    num_roots[circle] = circle.count_roots(**count_kwargs)
                 else:
                     add_root(root, multiplicity)
 
@@ -536,53 +539,43 @@ def find_roots_gen(
             # directly and disregard this contour (repeat its parent's subdivision).
             if contour.distance(root) < newton_step_tol:
                 # remove the contour and any relations
-                remove_siblings_children(contour)
-
-                # put the parent contour back into the list of contours to subdivide
-                # again
-                parent = contour._parent
-                contours.append(parent)
-
-                # do not use this contour again
-                failed_contours.append(contour)
+                reverse_subdivision(contour)
 
         # if we haven't found all the roots then subdivide further
         num_known_roots = sum(
-            [
-                int(round(multiplicity.real))
-                for root, multiplicity in zip(roots, multiplicities)
-                if contour.contains(root)
-            ]
+            multiplicity
+            for root, multiplicity in zip(roots, multiplicities)
+            if contour.contains(root)
         )
-        if contour._num_roots != num_known_roots and contour not in failed_contours:
+        if num_roots[contour] != num_known_roots and contour not in failed_contours:
             subdivide(contour)
-
-    # delete cache for original contour incase this contour is being reused
-    for segment in original_contour.segments:
-        segment._integralCache = {}
-        segment._trap_cache = {}
 
     result = RootResult(roots, multiplicities, original_contour)
     logger.info(
         "Completed rootfinding with "
         + str(f.calls)
-        + " evaluations of f at "
+        + " evaluations of the given analytic function at "
         + str(f.points)
         + " points\n"
         + str(result)
     )
 
-    num_found_roots = sum(
-        int(round(multiplicity.real))
-        for root, multiplicity in zip(roots, multiplicities)
+    num_found_roots = sum(multiplicities)
+    remaining_roots = num_roots[original_contour] - num_found_roots
+    yield RootFinderState(
+        roots, multiplicities, make_contour_data(contours, num_roots), remaining_roots
     )
-    remaining_roots = original_contour._num_roots - num_found_roots
-    yield roots, multiplicities, contours, remaining_roots
 
 
 @update_docstring(Parameters=FunctionDoc(find_roots_gen)["Parameters"])
 @functools.wraps(find_roots_gen, assigned=("__module__", "__name__"))
-def find_roots(original_contour, f, df=None, verbose=False, **kwargs):
+def find_roots(
+    original_contour: ContourABC,
+    f: AnalyticFunc,
+    df: Optional[AnalyticFunc] = None,
+    verbose: bool = False,
+    **kwargs,
+) -> RootResult:
     """
     Find all the roots of the complex analytic function f within the
     given contour.
@@ -598,6 +591,9 @@ def find_roots(original_contour, f, df=None, verbose=False, **kwargs):
     result : :class:`RootResult <cxroots.root_result.RootResult>`
         A container for the roots and their multiplicities.
     """
+    roots, multiplicities = [], []
+    root_finder = find_roots_gen(original_contour, f, df, **kwargs)
+
     if verbose:
         text_column = TextColumn("{task.description}")
         bar_column = BarColumn(bar_width=None)
@@ -605,25 +601,19 @@ def find_roots(original_contour, f, df=None, verbose=False, **kwargs):
             "{task.completed} of {task.total} roots found"
         )
         progress = Progress(text_column, bar_column, progress_text_column, expand=True)
-        # Set visible=False here so that we don't show the progress bar before the
-        # total number of roots in the contour have been determined
-        task = progress.add_task("Rootfinding", visible=False)
+        task = progress.add_task("Rootfinding")
         progress.start()
 
-    try:
-        root_finder = find_roots_gen(original_contour, f, df, **kwargs)
-        for roots, multiplicities, _, num_remaining_roots in root_finder:
-            if verbose:
-                num_found_roots = sum(
-                    int(round(multiplicity.real))
-                    for root, multiplicity in zip(roots, multiplicities)
-                )
+        try:
+            for roots, multiplicities, _, num_remaining_roots in root_finder:
+                num_found_roots = sum(multiplicities)
                 total_roots = num_found_roots + num_remaining_roots
-                progress.update(
-                    task, completed=num_found_roots, total=total_roots, visible=True
-                )
-    finally:
-        if verbose:
+                progress.update(task, completed=num_found_roots, total=total_roots)
+        finally:
             progress.stop()
+
+    else:
+        for roots, multiplicities, _, _ in root_finder:
+            pass
 
     return RootResult(roots, multiplicities, original_contour)